# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This service registry and controller new customer lab",
        "title": "platform-customer-registry",
        "contact": {
            "name": "Rafael Tomelin",
            "url": "https://local",
            "email": "contato@synera.com.br"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/labdestroy": {
            "get": {
                "description": "get all lab destroy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labDestroy"
                ],
                "summary": "get all lab destroy",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kubediscovery_platform-customer-registry_internal_core_entity.LabDestroyResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new lab destroy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labDestroy"
                ],
                "summary": "create a new lab destroy",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubediscovery_platform-customer-registry_internal_core_entity.LabDestroyResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/labdestroy/search": {
            "get": {
                "description": "get a lab destroy by email or project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labDestroy"
                ],
                "summary": "get a lab destroy by email or project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email@domain.com",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "project_name",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "string default",
                        "name": "available",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kubediscovery_platform-customer-registry_internal_core_entity.LabDestroyResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/labdestroy/{id}": {
            "get": {
                "description": "get a lab destroy by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labDestroy"
                ],
                "summary": "get a lab destroy by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "found",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubediscovery_platform-customer-registry_internal_core_entity.LabDestroyResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/registry": {
            "get": {
                "description": "get all customer registry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registry"
                ],
                "summary": "get all customer registry",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kubediscovery_platform-customer-registry_internal_core_entity.CustomerRegistryResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new registry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registry"
                ],
                "summary": "create a new customer registry",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubediscovery_platform-customer-registry_internal_core_entity.CustomerRegistryResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/registry/{id}": {
            "get": {
                "description": "get a customer registry by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registry"
                ],
                "summary": "get a customer registry by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "found",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubediscovery_platform-customer-registry_internal_core_entity.CustomerRegistryResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_kubediscovery_platform-customer-registry_internal_core_entity.CustomerRegistryResponse": {
            "type": "object",
            "required": [
                "email",
                "id",
                "project_name",
                "repository",
                "username"
            ],
            "properties": {
                "avaliable": {
                    "type": "boolean"
                },
                "create_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                },
                "repository": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_kubediscovery_platform-customer-registry_internal_core_entity.LabDestroyResponse": {
            "type": "object",
            "required": [
                "avaliable",
                "email",
                "id",
                "project_name",
                "repository",
                "username"
            ],
            "properties": {
                "avaliable": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                },
                "repository": {
                    "type": "string"
                },
                "target_destroy": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}